+++ elevenlabs.py
@@ -1,7 +1,9 @@
 import os
 from dotenv import load_dotenv
-load_dotenv()
 
+'''ElevenLabs API request information'''
+
+load_dotenv() # Load environment variables
 BASE_URL = "https://api.elevenlabs.io"
 HEADERS = {
     "xi-api-key": os.getenv('ELEVEN_LABS_API_KEY')