+++ voice.py
@@ -13,13 +13,12 @@ class Voice:
     def __init__(self, voice):
         '''
         Constructs a new voice
+
         Args: 
             voice (dict): ElevenLabs API response voice model
+
         Returns:
             void
-
-        Initalizes voice_id and name, as well as voice_lines, traits, and
-        nicknames if they are found
         '''
         
         self.voice_id = voice['voice_id']
@@ -31,19 +30,22 @@ class Voice:
             self.name = voice['name']
 
         self.image_path = character_settings.get(self.voice_id, {}).get('image_path')
+        self.image_position = character_settings.get(self.voice_id, {}).get('image_position')
 
     def CreateSpeech(self, text):
         '''
         Creates a synthesized speech instance through ElevenLabs
+
         Args:
             voice_id (string): ID of the voice in which speech should be created
             text (string): Text of the speech to be created
+
         Returns:
             Speech: New speech instance
         '''
-        if (len(text) < 3 or len(text) > 100):
+        if (len(text) < 3 or len(text) > 150):
             raise ValueError('Length of voice line must be between 3 and 100'
-            'characters')
+            ' characters')
         
         response = requests.post(el.BASE_URL + f"/v1/text-to-speech/{self.voice_id}/with-timestamps",
             headers=el.HEADERS,